requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
discord.py==2.4.0
disnake==2.9.2
docopt==0.6.2
env==0.1.0
executing==2.1.0
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.6
imageio-ffmpeg==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
py-cord==2.6.0
pycord==0.1.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
setuptools==74.1.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.10.0
